requirements.txt
ffastapi
uvicorn
jinja2
python-multipart
SQLAlchemy>=2.0
psycopg2-binary